+++ syncrclone/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '20201125.0.BETA'
+__version__ = '20201215.0.BETA'
 
 import time
 import io